move OperationPipelineExecutor from pipeline to pipeline/operations. And it's a parent interface for futur child pipeline backend as cpu or gpu. Execute and applyScheduling are now virtual
